Implemented global wind speed and direction for the 2D and cumulus clouds.
Adds a "Weather" section to the Sky3D node with properties to set both the wind speed (in m/s) and direction (in degrees, measured from the north). Changing these will set the appropriate direction and speed for both the 2D and cumulus clouds in the sky dome.
Also makes a few changes to the sky material shader to allow smooth transitions in the cloud animations when changing either the wind speed or direction. Before, clouds would "skip" whenever you made such changes.