Updated the 2D and cumulus clouds after discussions.
Fixes the issue with "jumps" by calculating the positions in GDScript, where we can keep state and take advantage of that so TIME wrapping after 3600 seconds no longer happens.
Also fixes the issue with "stutters" caused by positions with large values, that decrease the precision of the floating point numbers that are used.

These issues are fixed in a different way for both types of clouds.

For 2D clouds, the position, which translates to a UV position on the texture (actually twice to create some movement within the cloud like before), is simply wrapped between [0, 1) so values never get large and because we set this value in _process it will also nicely update each rendered frame.

For cumulus clouds, there does not seem to be any repeating pattern in how these clouds are drawn, so the above does not work. In stead we calculate two "layers" of cumulus clouds. We render the first until it moves away from the center too far, at which point it will slowly fade out, whilst the second layer fades in and starts moving from the origin. As soon as the first layer is invisible, we can reset its position back to zero and we will start moving it again as soon as the second layer moves too far away from the center and starts fading out. This process is repeated indefinitely.

Note that this commit is a proof of concept. There is more to fix before it could be merged.