Optimized the _process method to do as little as possible.
Refactored and optimized the _process method to only update clouds that are visible and to only set shader parameters if they actually changed. Whilst the code still is a couple of dozens of lines long, in most cases, only a very short path is actually covered. Most of the "extra" code comes for the edge case where we have to transition from one layer of cumulus clouds to the next so we can then "reset" the invisible layer back to the origin. In practice, this might not happen more than once an hour or so, depending on the configuration.